chore: Respository 타입값 수정

+++ BE/hp-domain/src/main/java/com/ssafy/hp/user/UserPillRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.hp.user;
 import com.ssafy.hp.user.domain.UserPill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserPillRepository extends JpaRepository<UserPill, Long> {
+public interface UserPillRepository extends JpaRepository<UserPill, Integer> {
 }
